ROS install apt source fix

+++ install_ROS2.sh
@@ -15,12 +15,17 @@ source /etc/os-release
 sudo apt install software-properties-common -y
 sudo add-apt-repository universe -y
 sudo apt update && sudo apt install curl -y
-ROS_APT_SOURCE_VERSION=$(curl -s https://api.github.com/repos/ros-infrastructure/ros-apt-source/releases/latest | grep -F "tag_name" | awk -F\" '{print $4}')
-if [ -z "$ROS_APT_SOURCE_VERSION" ]; then
-	ROS_APT_SOURCE_VERSION="1.1.0"
+
+# check if ros2-apt-source is already installed, if not, install it
+if ! dpkg -s ros2-apt-source >/dev/null 2>&1; then
+
+	ROS_APT_SOURCE_VERSION=$(curl -s https://api.github.com/repos/ros-infrastructure/ros-apt-source/releases/latest | grep -F "tag_name" | awk -F\" '{print $4}')
+	if [ -z "$ROS_APT_SOURCE_VERSION" ]; then
+		ROS_APT_SOURCE_VERSION="1.2.0" # when curl fails, use a fallback version, latest as of april 2026
+	fi
+	curl -L -o /tmp/ros2-apt-source.deb "https://github.com/ros-infrastructure/ros-apt-source/releases/download/${ROS_APT_SOURCE_VERSION}/ros2-apt-source_${ROS_APT_SOURCE_VERSION}.$(. /etc/os-release && echo $VERSION_CODENAME)_all.deb"
+	sudo apt install /tmp/ros2-apt-source.deb
 fi
-curl -L -o /tmp/ros2-apt-source.deb "https://github.com/ros-infrastructure/ros-apt-source/releases/download/${ROS_APT_SOURCE_VERSION}/ros2-apt-source_${ROS_APT_SOURCE_VERSION}.$(. /etc/os-release && echo $VERSION_CODENAME)_all.deb"
-sudo apt install /tmp/ros2-apt-source.deb
 sudo apt update
 
 # select ros2 type based on UBUNTU_CODENAME
